Extract ReaderSidebarTagsSection to a separate view
