+++ WordPress/Classes/ViewRelated/Reader/ReaderSidebarViewController.swift
@@ -160,98 +160,8 @@ struct ReaderSidebarView: View {
     }
 }
 
-private struct ReaderSidebarSubscriptionsSection: View {
-    let viewModel: ReaderSidebarViewModel
-
-    @FetchRequest(
-        sortDescriptors: [SortDescriptor(\.title, order: .forward)],
-        predicate: NSPredicate(format: "following = YES")
-    )
-    private var subscriptions: FetchedResults<ReaderSiteTopic>
-
-    var body: some View {
-        Label(Strings.allSubscriptions, systemImage: "checkmark.rectangle.stack")
-            .tag(ReaderSidebarItem.allSubscriptions)
-
-        ForEach(subscriptions, id: \.self) { site in
-            Label {
-                Text(site.title)
-            } icon: {
-                SiteIconView(viewModel: SiteIconViewModel(readerSiteTopic: site, size: .small))
-                    .frame(width: 28, height: 28)
-            }
-            .lineLimit(1)
-            .tag(ReaderSidebarItem.subscription(TaggedManagedObjectID(site)))
-            .swipeActions(edge: .trailing) {
-                Button(SharedStrings.Reader.unfollow, role: .destructive) {
-                    ReaderSubscriptionHelper().unfollow(site)
-                }.tint(.red)
-            }
-        }
-        .onDelete(perform: delete)
-    }
-
-    func delete(at offsets: IndexSet) {
-        let sites = offsets.map { subscriptions[$0] }
-        for site in sites {
-            ReaderSubscriptionHelper().unfollow(site)
-        }
-    }
-}
-
-private struct ReaderSidebarTagsSection: View {
-    let viewModel: ReaderSidebarViewModel
-
-    @FetchRequest(
-        sortDescriptors: [SortDescriptor(\.title, order: .forward)],
-        predicate: ReaderSidebarTagsSection.predicate
-    )
-    private var tags: FetchedResults<ReaderTagTopic>
-
-    static let predicate = NSPredicate(format: "following == YES AND showInMenu == YES AND type == 'tag'")
-
-    var body: some View {
-        ForEach(tags, id: \.self) { tag in
-            Label {
-                Text(tag.title)
-                    .lineLimit(1)
-            } icon: {
-                Image(systemName: "tag")
-                    .foregroundStyle(.secondary)
-            }
-            .tag(ReaderSidebarItem.tag(TaggedManagedObjectID(tag)))
-            .swipeActions(edge: .trailing) {
-                Button(SharedStrings.Reader.unfollow, role: .destructive) {
-                    // TODO: (wpsidebar) implement
-                }.tint(.red)
-            }
-        }
-        .onDelete(perform: delete)
-
-        Button {
-            // TODO: (wpsidebar) implement
-        } label: {
-            Label(Strings.addTag, systemImage: "plus.circle")
-        }
-
-        Button {
-            viewModel.navigate(.discoverTags)
-        } label: {
-            Label(Strings.discoverTags, systemImage: "sparkle.magnifyingglass")
-        }
-    }
-
-    func delete(at offsets: IndexSet) {
-        // TODO: (wpsidebar) implement
-    }
-}
-
 private struct Strings {
     static let reader = NSLocalizedString("reader.sidebar.navigationTitle", value: "Reader", comment: "Reader sidebar title")
-    static let allSubscriptions = NSLocalizedString("reader.sidebar.allSubscriptions", value: "All Subscriptions", comment: "Reader sidebar button title")
-    static let addSubscription = NSLocalizedString("reader.sidebar.addSubscription", value: "Add Subscription", comment: "Reader sidebar button title")
     static let subscriptions = NSLocalizedString("reader.sidebar.section.subscriptions.tTitle", value: "Subscriptions", comment: "Reader sidebar section title")
     static let tags = NSLocalizedString("reader.sidebar.section.tags.title", value: "Tags", comment: "Reader sidebar section title")
-    static let addTag = NSLocalizedString("reader.sidebar.section.tags.addTag", value: "Add tag", comment: "Reader sidebar button")
-    static let discoverTags = NSLocalizedString("reader.sidebar.section.tags.discoverTags", value: "Discover More Tags", comment: "Reader sidebar button")
 }
